optimizing scrolling
adjusting the listview to show the first post item on load/start
when the actiobar is clicked the listview scrolls to the first item.
If the current API Level is greater than 11, then the view should align perfectly
by smooth scrolling, otherwise smoothscrolling is turned off.

If the user scrolls enough, so that the search view appears, then the keyboard is shown.
If the search view disappears during scrolling, then the keyboard is removed.